# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Aisync API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
  /** @minLength 2 */
  businessName: string;
  industry: string;
  /** @minLength 5 */
  phone: string;
  /**
     * What the client wants the voice agent to do
     * @minLength 10
     */
  requirements: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  client: 'client',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
}

export interface DemoRequestInput {
  name: string;
  businessName: string;
  email: string;
  phone: string;
  industry: string;
  /** @nullable */
  message?: string | null;
}

export interface DemoRequest {
  id: number;
  name: string;
  businessName: string;
  email: string;
  phone: string;
  industry: string;
  /** @nullable */
  message?: string | null;
  createdAt: string;
}

export interface ClientInput {
  name: string;
  email: string;
  businessName: string;
  industry: string;
  phone: string;
  /** @nullable */
  notes?: string | null;
  password?: string;
}

export interface ClientUpdate {
  businessName?: string;
  industry?: string;
  phone?: string;
  /** @nullable */
  notes?: string | null;
}

export interface ClientWithUser {
  id: number;
  userId: number;
  businessName: string;
  industry: string;
  phone: string;
  email: string;
  /** @nullable */
  notes?: string | null;
  userName?: string;
  userEmail?: string;
}

export type ProjectInputStatus = typeof ProjectInputStatus[keyof typeof ProjectInputStatus];


export const ProjectInputStatus = {
  onboarding: 'onboarding',
  build_phase: 'build_phase',
  demo_phase: 'demo_phase',
  testing: 'testing',
  completed: 'completed',
} as const;

export interface ProjectInput {
  clientId: number;
  projectName: string;
  description: string;
  status: ProjectInputStatus;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  expectedCompletionDate?: string | null;
  /** @nullable */
  latestUpdate?: string | null;
}

export type ProjectUpdateStatus = typeof ProjectUpdateStatus[keyof typeof ProjectUpdateStatus];


export const ProjectUpdateStatus = {
  onboarding: 'onboarding',
  build_phase: 'build_phase',
  demo_phase: 'demo_phase',
  testing: 'testing',
  completed: 'completed',
} as const;

export interface ProjectUpdate {
  projectName?: string;
  description?: string;
  status?: ProjectUpdateStatus;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  expectedCompletionDate?: string | null;
  /** @nullable */
  latestUpdate?: string | null;
}

export interface Employee {
  id: number;
  name: string;
  role: string;
  email: string;
}

export type ProjectWithDetailsStatus = typeof ProjectWithDetailsStatus[keyof typeof ProjectWithDetailsStatus];


export const ProjectWithDetailsStatus = {
  onboarding: 'onboarding',
  build_phase: 'build_phase',
  demo_phase: 'demo_phase',
  testing: 'testing',
  completed: 'completed',
} as const;

export interface ProjectWithDetails {
  id: number;
  clientId: number;
  projectName: string;
  description: string;
  status: ProjectWithDetailsStatus;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  expectedCompletionDate?: string | null;
  /** @nullable */
  latestUpdate?: string | null;
  clientBusinessName?: string;
  employees?: Employee[];
}

export interface AssignmentInput {
  employeeId: number;
}

export interface EmployeeInput {
  name: string;
  role: string;
  email: string;
}

export interface EmployeeUpdate {
  name?: string;
  role?: string;
  email?: string;
}

export type InvoiceInputStatus = typeof InvoiceInputStatus[keyof typeof InvoiceInputStatus];


export const InvoiceInputStatus = {
  paid: 'paid',
  partial: 'partial',
  due: 'due',
  overdue: 'overdue',
} as const;

export interface InvoiceInput {
  clientId: number;
  /** @nullable */
  projectId?: number | null;
  invoiceNumber: string;
  totalAmount: number;
  paidAmount: number;
  dueAmount: number;
  status: InvoiceInputStatus;
  dueDate: string;
  /** @nullable */
  notes?: string | null;
}

export type InvoiceUpdateStatus = typeof InvoiceUpdateStatus[keyof typeof InvoiceUpdateStatus];


export const InvoiceUpdateStatus = {
  paid: 'paid',
  partial: 'partial',
  due: 'due',
  overdue: 'overdue',
} as const;

export interface InvoiceUpdate {
  invoiceNumber?: string;
  totalAmount?: number;
  paidAmount?: number;
  dueAmount?: number;
  status?: InvoiceUpdateStatus;
  dueDate?: string;
  /** @nullable */
  notes?: string | null;
}

export type InvoiceStatus = typeof InvoiceStatus[keyof typeof InvoiceStatus];


export const InvoiceStatus = {
  paid: 'paid',
  partial: 'partial',
  due: 'due',
  overdue: 'overdue',
} as const;

export interface Invoice {
  id: number;
  clientId: number;
  /** @nullable */
  projectId?: number | null;
  invoiceNumber: string;
  totalAmount: number;
  paidAmount: number;
  dueAmount: number;
  status: InvoiceStatus;
  dueDate: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  /** @nullable */
  clientBusinessName?: string | null;
}

export interface FileUploadInput {
  fileName: string;
  /** @nullable */
  projectId?: number | null;
}

export interface UploadedFile {
  id: number;
  clientId: number;
  /** @nullable */
  projectId?: number | null;
  fileName: string;
  filePath: string;
  uploadedAt: string;
}

export type FeatureRequestInputPriority = typeof FeatureRequestInputPriority[keyof typeof FeatureRequestInputPriority];


export const FeatureRequestInputPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface FeatureRequestInput {
  /** @nullable */
  projectId?: number | null;
  title: string;
  description: string;
  priority: FeatureRequestInputPriority;
}

export type FeatureRequestUpdateStatus = typeof FeatureRequestUpdateStatus[keyof typeof FeatureRequestUpdateStatus];


export const FeatureRequestUpdateStatus = {
  new: 'new',
  in_review: 'in_review',
  approved: 'approved',
  completed: 'completed',
  rejected: 'rejected',
} as const;

export type FeatureRequestUpdatePriority = typeof FeatureRequestUpdatePriority[keyof typeof FeatureRequestUpdatePriority];


export const FeatureRequestUpdatePriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface FeatureRequestUpdate {
  status?: FeatureRequestUpdateStatus;
  title?: string;
  description?: string;
  priority?: FeatureRequestUpdatePriority;
}

export type FeatureRequestPriority = typeof FeatureRequestPriority[keyof typeof FeatureRequestPriority];


export const FeatureRequestPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export type FeatureRequestStatus = typeof FeatureRequestStatus[keyof typeof FeatureRequestStatus];


export const FeatureRequestStatus = {
  new: 'new',
  in_review: 'in_review',
  approved: 'approved',
  completed: 'completed',
  rejected: 'rejected',
} as const;

export interface FeatureRequest {
  id: number;
  clientId: number;
  /** @nullable */
  projectId?: number | null;
  title: string;
  description: string;
  priority: FeatureRequestPriority;
  status: FeatureRequestStatus;
  createdAt: string;
  /** @nullable */
  clientBusinessName?: string | null;
}

export interface AdminSummary {
  totalClients: number;
  activeProjects: number;
  completedProjects: number;
  totalPaid: number;
  totalDue: number;
  pendingFeatureRequests: number;
}