# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Aisync API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
  /** @minLength 2 */
  businessName: string;
  industry: string;
  /** @minLength 5 */
  phone: string;
  /**
     * What the client wants the voice agent to do
     * @minLength 10
     */
  requirements: string;
}